refactor: Piece 패키지 분리 및 메소드 정리
