+++ src/main/java/chess/domain/piece/fixedmovablepiece/Knight.java
@@ -1,5 +1,7 @@
-package chess.domain.piece;
+package chess.domain.piece.fixedmovablepiece;
 
+import chess.domain.piece.Color;
+import chess.domain.piece.PieceName;
 import chess.domain.position.Direction;
 import chess.domain.position.Position;
 import java.util.List;
@@ -7,7 +9,7 @@ import java.util.Map;
 
 public class Knight extends FixedMovablePiece {
 
-    public final static List<Direction> DIRECTIONS = List.of(
+    private final static List<Direction> DIRECTIONS = List.of(
             Direction.SSE, Direction.SSW, Direction.NNE, Direction.NNW,
             Direction.EES, Direction.EEN, Direction.WWS, Direction.WWN);
