+++ src/main/java/chess/domain/piece/straightmovablepiece/Bishop.java
@@ -1,5 +1,7 @@
-package chess.domain.piece;
+package chess.domain.piece.straightmovablepiece;
 
+import chess.domain.piece.Color;
+import chess.domain.piece.PieceName;
 import chess.domain.position.Direction;
 import chess.domain.position.Position;
 import java.util.List;
@@ -10,10 +12,8 @@ public class Bishop extends StraightMovablePiece {
     private final static List<Direction> DIRECTIONS = List.of(
             Direction.NORTH_EAST, Direction.NORTH_WEST, Direction.SOUTH_EAST, Direction.SOUTH_WEST);
 
-    public static List<Position> BLACK_INIT_LOCATIONS = List.of(
-            Position.of("c8"), Position.of("f8"));
-    public static List<Position> WHITE_INIT_LOCATIONS = List.of(
-            Position.of("c1"), Position.of("f1"));
+    public static List<Position> BLACK_INIT_LOCATIONS = List.of(Position.of("c8"), Position.of("f8"));
+    public static List<Position> WHITE_INIT_LOCATIONS = List.of(Position.of("c1"), Position.of("f1"));
 
     public Bishop(Color color) {
         super(color, PieceName.BISHOP);
