+++ src/main/java/chess/domain/piece/straightmovablepiece/StraightMovablePiece.java
@@ -1,5 +1,8 @@
-package chess.domain.piece;
+package chess.domain.piece.straightmovablepiece;
 
+import chess.domain.piece.Color;
+import chess.domain.piece.Piece;
+import chess.domain.piece.PieceName;
 import chess.domain.position.Direction;
 import chess.domain.position.Position;
 import java.util.ArrayList;
@@ -7,16 +10,16 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public abstract class FixedMovablePiece extends Piece {
+public abstract class StraightMovablePiece extends Piece {
 
-    protected FixedMovablePiece(Color color, PieceName name) {
+    protected StraightMovablePiece(Color color, PieceName name) {
         super(color, name);
     }
 
     @Override
     public abstract Map<Direction, List<Position>> getMovablePositions(Position position);
 
-    protected Map<Direction, List<Position>> getMovablePositionsByDirections(Position position, List<Direction> directions) {
+    protected final Map<Direction, List<Position>> getMovablePositionsByDirections(Position position, List<Direction> directions) {
         Map<Direction, List<Position>> movable = new HashMap<>();
         for (Direction direction : directions) {
             movable.put(direction, new ArrayList<>());
@@ -25,17 +28,13 @@ public abstract class FixedMovablePiece extends Piece {
         return movable;
     }
 
-    private void putMovablePositionsByDirection(Map<Direction, List<Position>> movable, Position position,
+    private final void putMovablePositionsByDirection(Map<Direction, List<Position>> movable, Position position,
                                                 Direction direction) {
         Position nextPosition = position.toDirection(direction);
         if (nextPosition == position) {
             return;
         }
         movable.get(direction).add(nextPosition);
-    }
-
-    @Override
-    public double getPoint() {
-        return 0;
+        putMovablePositionsByDirection(movable, nextPosition, direction);
     }
 }
