+++ src/test/java/chess/domain/ChessBoardTest.java
@@ -4,18 +4,18 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import chess.domain.piece.Bishop;
+import chess.domain.piece.straightmovablepiece.Bishop;
 import chess.domain.piece.Color;
 import chess.domain.piece.EmptyPiece;
-import chess.domain.piece.King;
-import chess.domain.piece.Knight;
+import chess.domain.piece.fixedmovablepiece.King;
+import chess.domain.piece.fixedmovablepiece.Knight;
 import chess.domain.piece.generator.NormalPiecesGenerator;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceName;
 import chess.domain.piece.generator.PiecesGenerator;
-import chess.domain.piece.Queen;
-import chess.domain.piece.Rook;
+import chess.domain.piece.straightmovablepiece.Queen;
+import chess.domain.piece.straightmovablepiece.Rook;
 import chess.domain.position.Column;
 import chess.domain.position.Position;
 import java.util.HashMap;
