+++ src/test/java/chess/domain/piece/RookTest.java
@@ -2,6 +2,7 @@ package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import chess.domain.piece.straightmovablepiece.Rook;
 import chess.domain.position.Direction;
 import chess.domain.position.Position;
 import java.util.HashMap;
